+++ lotfresh-front/src/api/cart/cart.ts
@@ -1,5 +1,5 @@
 import { defaultInstance, cartInstance } from "../utils";
-import type { CartCreateDto, OrderSheetInfo } from "@/interface/cartInterface";
+import type { CartCreateDto } from "@/interface/cartInterface";
 
 export const createCart = async (
   cartCreateDto: CartCreateDto,
@@ -9,19 +9,4 @@ export const createCart = async (
     headers: { Authorization: `Bearer ${accessToken}` },
   });
   return response;
-};
-
-export const addOrderheetInfos = async (
-  orderSheetInfos: OrderSheetInfo[],
-  province: string | null | undefined,
-  productId: number,
-  accessToken: string | null
-): Promise<any> => {
-  const response = await cartInstance.post(
-    `/carts/province/${province}/products`,
-    orderSheetInfos,
-    {
-      headers: { Authorization: `Bearer ${accessToken}` },
-    }
-  );
 };
