+++ lotfresh-front/src/interface/memberInterface.ts
@@ -11,3 +11,17 @@ export interface CreateMemberDto {
   zipCode: string;
   addressDetail: string | null | undefined;
 }
+
+export interface OrderSheetItem {
+  productId: number;
+  originalPrice: number;
+  discountedPrice: number;
+  productStock: number;
+  productName: string;
+  productThumbnail: string;
+}
+
+export interface OrderSheetList {
+  orderSheetItems: OrderSheetItem[] | null;
+  isFromCart: boolean;
+}
